Refactor DebugTensorValueComponent into separate folder
