+++ tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/debug_tensor_value/BUILD
@@ -1,0 +1,50 @@
+package(default_visibility = ["//tensorboard:internal"])
+
+load("@npm_angular_bazel//:index.bzl", "ng_module")
+load("//tensorboard/defs:defs.bzl", "tf_ts_library")
+
+licenses(["notice"])  # Apache 2.0
+
+ng_module(
+    name = "debug_tensor_value",
+    srcs = [
+        "debug_tensor_value_component.ts",
+        "debug_tensor_value_module.ts",
+    ],
+    deps = [
+        "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/store:debug_tensor_value",
+        "@npm//@angular/common",
+        "@npm//@angular/core",
+    ],
+)
+
+# tf_ts_library(
+#     name = "alerts_container_test_lib",
+#     testonly = True,
+#     srcs = [
+#         "alerts_container_test.ts",
+#     ],
+#     tsconfig = "//:tsconfig-test",
+#     deps = [
+#         ":alerts",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin:debugger_v2",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/actions",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/store",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/store:types",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/testing",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/execution_data",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/graph_executions",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/inactive",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/source_files",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/stack_trace",
+#         "//tensorboard/plugins/debugger_v2/tf_debugger_v2_plugin/views/timeline",
+#         "//tensorboard/webapp/angular:expect_angular_core_testing",
+#         "//tensorboard/webapp/angular:expect_ngrx_store_testing",
+#         "@npm//@angular/common",
+#         "@npm//@angular/compiler",
+#         "@npm//@angular/core",
+#         "@npm//@angular/platform-browser",
+#         "@npm//@ngrx/store",
+#         "@npm//@types/jasmine",
+#     ],
+# )
